requirements.txt
streamlitpandasopenpyxlrequestsPillowbeautifulsoup4